pyproject.toml
[tool.poetry]
name = "edgewise"
version = "0.1.0"
description = "An Active Record ORM for EdgeDB"
authors = ["Tyler Matteson <tyler@agritheory.com>"]
readme = "readme.md"
repository = "https://github.com/agritheory/edgewise"
license = "Apache-2.0"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python",
]

[tool.poetry.dependencies]
python = "^3.8"
edgedb = "^0.7.1"
attrs = "^19.3.0"
wrapt = "^1.12.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
mimesis = "^4.0.0"
pytest-asyncio = "^0.10.0"
pytest-cov = "^2.8.1"
pre-commit = "^2.2.0"


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
